feat: add login functionality to menu screen
This commit adds the functionality to login players from the menu screen. It sends a POST request to the login endpoint with the provided username and password. If the login is successful, it prints the token. If there is an error, it provides appropriate error messages based on the response status code.